+++ src/components/editors/gesture/index.ts
@@ -1,0 +1,2 @@
+export { GestureHoldActivatorEditor } from "./GestureHoldActivatorEditor";
+export { GesturePathEditor } from "./GesturePathEditor";