+++ src/components/editors/index.ts
@@ -1,11 +1,1 @@
-export { ActivatorEditor } from "./ActivatorEditor";
-export { ComboActivatorEditor } from "./ComboActivatorEditor";
-export { CustomKeyEditor } from "./CustomKeyEditor";
 export { EditorHost } from "./EditorHost";
-export { KeyChordModal } from "./KeyChordModal";
-export { MacroDelayPrompt } from "./MacroDelayPrompt";
-export { MacroEditor } from "./MacroEditor";
-export { MacroKeyPrompt } from "./MacroKeyPrompt";
-export { OpenAppEditor } from "./OpenAppEditor";
-export { OpenAppList } from "./OpenAppList";
-export { OpenAppPickerContent } from "./OpenAppPickerContent";